merge Bug fix by Bruno-w (https://github.com/Bruno-w): RTC alarm wake-up
Because of the S3C2410 bug (INT_SRCPND_RTC not set on alarm wake-up), the
firmware need to check on power-on if an alarm is triggered by comparing the
alarm date to the system date.  But, It failed because the firmware read a
system date that is 1 second behind the alarm date, while the cpu was waked-up
by an alarm.  This is because x49gp uses 2 different functions to read host
system time in 2 key routines, leading to a 1 second gap when used consecutively.
This fix solved this issue.